Some figures
Two figures from the intro frequency sim.
Still not really based closely on real data or noise estimates.
propeachfreq - takes the frequency with the highest power from this simulation and plots the histogram of all the simulations. Hope for a peak at closest to the simulated frequency.
timecoursepower - plots the pdif timecourse at the sampled delays with the addition of noise. And then the power for each frequency. This is just one example simulation.